central room implemented
the large central room and its connected halls have been fully implemented. With that the majority of the grey boxing is complete and the map resembles much of the original paper design. Further details will be added from here, like railing, various entities, lighting and of course texturing. But for the most part, the skeleton of the level is complete. Additional notes, I dealt with some level geometry that was giving me trouble, cleaned up certain brushes and ugly vertices, and also switched to working mostly in units of 8